AI-Portfolio-Optimizer/backend/requirements.txt
fastapi
uvicorn
pandas
yfinance
PyPortfolioOpt
python-dotenv
cachetools
